Added PDF with the schematic and PCB design
Two of the components are underneath the arduino nano. I did this on purpose because I plan on using female headers to connect the nano to the board so that it can be easily removed. This will prop up the board far enough to have a small resistor and capacitor underneath.